requirements.txt
streamlit
plotly
pandas
scikit-learn
os-sys
ucimlrepo
ydata-profiling
seaborn
matplotlib.pyplot
transformers